Added the ability read configuration from a config file.
This is required because in certain scenarios the length of the command
line arguments may hit a limit. Therefore by writing those command line
arguments to a file and providing DbMigrator with the path to the file
using a -ConfigFile command line argument which will read the arguments
from the file rather than needing to pass the arguments in via the
command line.

Updated the readme file.