Moved the check to see whether the user has requested help
Moved the check to see if the "-help" command line parameter has been
passed to ensure that no other parameters are required. The check for
the help parameter is now the first check that gets done and since the
application returns after displaying the help output no other parameters
are required.